API look works now, need to tweak for overtimes
