+++ .gitignore
@@ -2,3 +2,5 @@ dist
 
 node_modules
 app.zip
+.vscode/settings.json
+.vscode/settings.json
