+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
     "tfvc.location": "C:\\Program Files (x86)\\Microsoft Visual Studio\\2019\\Community\\Common7\\IDE\\CommonExtensions\\Microsoft\\TeamFoundation\\Team Explorer\\TF.exe"
+    ,"prettier.printWidth": 140
 }
