+++ src/scripts/app.ts
@@ -1,9 +1,6 @@
-import { BasketballGame } from "./basketballGame";
 import { Scheduler } from "./helpers/scheduler";
-import { Helpers } from "./helpers/helpers";
 
 const moment = require("moment");
-
 let jsdom = require("jsdom");
 let $ = require("jquery")(new jsdom.JSDOM().window);
 const rp = require("request-promise");
@@ -12,4 +9,4 @@ const fs = require("fs");
 const scheduler = require('node-schedule');
 
 // Schedule all games for the day
-Scheduler.scheduleAllGames();
+Scheduler.scheduleAllScraperGames();
