+++ src/scripts/test.ts
@@ -1,5 +1,9 @@
-import { BasketballGame } from "./basketballGame";
+import { BasketballGameScraper } from "./basketballGameScraper";
 import { Scheduler } from "./helpers/scheduler";
 
+import { BasketballGame } from "./basketballGame";
+
 //Add any tests here
-Scheduler.scheduleNow();
+//Scheduler.scheduleAllAPIGames();
+
+Scheduler.scheduleAllAPIGames();
